Moving parts of `System.MathF` to be shared with CoreRT.
