+++ src/classlibnative/inc/floatsingle.h
@@ -20,11 +20,11 @@ public:
     FCDECL1_V(static float, Cosh, float x);
     FCDECL1_V(static float, Exp, float x);
     FCDECL1_V(static float, Floor, float x);
+    FCDECL2_VV(static float, FMod, float x, float y);
     FCDECL1_V(static float, Log, float x);
     FCDECL1_V(static float, Log10, float x);
-    FCDECL1(static float, ModF, float* iptr);
+    FCDECL2_VV(static float, ModF, float x, float* intptr);
     FCDECL2_VV(static float, Pow, float x, float y);
-    FCDECL1_V(static float, Round, float x);
     FCDECL1_V(static float, Sin, float x);
     FCDECL1_V(static float, Sinh, float x);
     FCDECL1_V(static float, Sqrt, float x);
